[INTERNAL] Refactored a few modules
