+++ lib/cli/commands/init.js
@@ -7,7 +7,7 @@ const initCommand = {
 
 initCommand.lazyRequireDependencies = async function() {
 	return {
-		validate: require("../../utils/validate"),
+		fsHelper: require("../../utils/fsHelper"),
 		init: require("../../init/init"),
 		promisify: require("util").promisify,
 		path: require("path"),
@@ -17,11 +17,11 @@ initCommand.lazyRequireDependencies = async function() {
 };
 
 initCommand.handler = async function() {
-	const {validate, init, promisify, path, fs, jsYaml} = initCommand.lazyRequireDependencies();
+	const {fsHelper, init, promisify, path, fs, jsYaml} = initCommand.lazyRequireDependencies();
 	const writeFile = promisify(fs.writeFile);
 	try {
 		const yamlPath = path.resolve("./ui5.yaml");
-		if (await validate.exists(yamlPath)) {
+		if (await fsHelper.exists(yamlPath)) {
 			throw new Error("Initialization not possible: ui5.yaml already exists");
 		}
