+++ lib/cli/commands/serve.js
@@ -68,13 +68,16 @@ serve.handler = function(argv = {}) {
 			cert: argv.h2 ? argv.cert : undefined,
 			key: argv.h2 ? argv.key : undefined
 		};
-		return !serverConfig.h2
-			? {serverConfig, tree}
-			: ui5Server.sslUtil.getSslCertificate(serverConfig.key, serverConfig.cert).then(({key, cert}) => {
+
+		if (!serverConfig.h2) {
+			return {serverConfig, tree};
+		} else {
+			return ui5Server.sslUtil.getSslCertificate(serverConfig.key, serverConfig.cert).then(({key, cert}) => {
 				serverConfig.key = key;
 				serverConfig.cert = cert;
 				return {serverConfig, tree};
 			});
+		}
 	}).then(({serverConfig, tree}) => {
 		return server.serve(tree, serverConfig).then(function({h2, port}) {
 			const protocol = h2 ? "https" : "http";
