+++ lib/init/init.js
@@ -2,7 +2,7 @@ const {promisify} = require("util");
 const path = require("path");
 const fs = require("fs");
 const readFile = promisify(fs.readFile);
-const validate = require("../utils/validate");
+const fsHelper = require("../utils/fsHelper");
 
 /**
  * Reads the package.json file and returns its content
@@ -87,7 +87,7 @@ async function init({cwd = "./"} = {}) {
 		throw new Error("Initialization not possible: Missing 'name' in package.json");
 	}
 
-	const [hasWebapp, hasSrc, hasTest] = await validate.pathsExist(["webapp", "src", "test"], cwd);
+	const [hasWebapp, hasSrc, hasTest] = await fsHelper.pathsExist(["webapp", "src", "test"], cwd);
 	projectConfig.type = getProjectType(hasWebapp, hasSrc, hasTest);
 
 	return projectConfig;
