+++ lib/utils/fsHelper.js
